+++ software/tracksight/frontend/src/packages/lib/index.ts
@@ -1,1 +1,2 @@
 export * from "./utils";
+export * from "./dashboardService";
